+++ src/main/se/kth/iv1350/integration/dto/ItemDTO.java
@@ -1,8 +1,6 @@
-package se.kth.iv1350.integration;
+package se.kth.iv1350.integration.dto;
 
 import se.kth.iv1350.model.Amount;
-
-import java.util.Objects;
 
 //TODO @Override toString()?
 // TODO Move to integration layer?
