+++ src/test/se/kth/iv1350/model/CustomerTest.java
@@ -2,7 +2,7 @@ package se.kth.iv1350.model;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import se.kth.iv1350.integration.CustomerDTO;
+import se.kth.iv1350.integration.dto.CustomerDTO;
 import se.kth.iv1350.integration.pricing.CustomerType;
 
 import static org.junit.jupiter.api.Assertions.*;
